+++ tests/test_core.py
@@ -43,6 +43,10 @@ def test_default_valdation_msg():
         arg.validated(1)
     assert 'Validator <lambda>(1) is not True' in str(excinfo)
 
+def test_conversion_to_str():
+    arg = Arg(str)
+    assert arg.validated(42) == '42'
+
 # Parser tests
 
 @mock.patch('webargs.core.Parser.parse_json')
@@ -50,7 +54,7 @@ def test_parse_json_called_by_parse_arg(parse_json, request):
     arg = Arg()
     p = Parser()
     p.parse_arg('foo', arg, request)
-    assert parse_json.called_once_with(request, 'foo')
+    assert parse_json.called
 
 
 @mock.patch('webargs.core.Parser.parse_querystring')
@@ -58,21 +62,14 @@ def test_parse_querystring_called_by_parse_arg(parse_querystring, request):
     arg = Arg()
     p = Parser()
     p.parse_arg('foo', arg, request)
-    assert parse_querystring.called_once_with(request, 'foo')
+    assert parse_querystring.called
 
 @mock.patch('webargs.core.Parser.parse_form')
 def test_parse_form_called_by_parse_arg(parse_form, request):
     arg = Arg()
     p = Parser()
     p.parse_arg('foo', arg, request)
-    assert parse_form.called_once_with(request, 'foo')
-
-@mock.patch('webargs.core.Parser.parse_headers')
-def test_parse_headers_called_by_parse_arg(parse_headers, request):
-    arg = Arg()
-    p = Parser()
-    p.parse_arg('foo', arg, request)
-    assert parse_headers.called_once_with(request, 'foo')
+    assert parse_form.called
 
 @mock.patch('webargs.core.Parser.parse_json')
 def test_parse_json_not_called_when_json_not_a_target(parse_json, request):
@@ -81,12 +78,30 @@ def test_parse_json_not_called_when_json_not_a_target(parse_json, request):
     p.parse_arg('foo', arg, request, targets=('form', 'querystring'))
     assert parse_json.call_count == 0
 
+@mock.patch('webargs.core.Parser.parse_headers')
+def test_parse_headers_called_when_headers_is_a_target(parse_headers, request):
+    arg = Arg()
+    p = Parser()
+    p.parse_arg('foo', arg, request)
+    assert parse_headers.call_count == 0
+    p.parse_arg('foo', arg, request, targets=('headers',))
+    assert parse_headers.called
+
+@mock.patch('webargs.core.Parser.parse_cookies')
+def test_parse_cookies_called_when_cookies_is_a_target(parse_cookies, request):
+    arg = Arg()
+    p = Parser()
+    p.parse_arg('foo', arg, request)
+    assert parse_cookies.call_count == 0
+    p.parse_arg('foo', arg, request, targets=('cookies',))
+    assert parse_cookies.called
+
 @mock.patch('webargs.core.Parser.fallback')
 def test_fallback_used_if_all_other_functions_return_none(fallback, request):
     arg = Arg()
     p = Parser()
     p.parse_arg('foo', arg, request)
-    assert fallback.called_once_with(request, 'foo')
+    assert fallback.called
 
 @mock.patch('webargs.core.Parser.parse_json')
 def test_parse(parse_json, request):
@@ -123,3 +138,26 @@ def test_value_error_raised_if_invalid_target(request):
     with pytest.raises(ValueError) as excinfo:
         p.parse_arg('foo', arg, request, targets=('invalidtarget', 'headers'))
     assert 'Invalid targets arguments: {0}'.format(['invalidtarget']) in str(excinfo)
+
+@mock.patch('webargs.core.Parser.parse_json')
+def test_conversion(parse_json, request):
+    parse_json.return_value = 42
+    arg = Arg(str)
+    assert Parser().parse_arg('foo', arg, request, targets=('json',)) == '42'
+
+@mock.patch('webargs.core.Parser.handle_error')
+@mock.patch('webargs.core.Parser.parse_json')
+def test_handle_error_called_when_parsing_raises_error(parse_json, handle_error, request):
+    val_err = ValidationError('error occurred')
+    parse_json.side_effect = val_err
+    p = Parser()
+    p.parse({'foo': Arg()}, request, targets=('json',))
+    assert handle_error.called
+    parse_json.side_effect = Exception('generic exception')
+    p.parse({'foo': Arg()}, request, targets=('json',))
+    assert handle_error.call_count == 2
+
+def test_handle_error_reraises_errors():
+    p = Parser()
+    with pytest.raises(ValidationError):
+        p.handle_error(ValidationError('error raised'))